requirements.txt
astropy==5.1
matplotlib==3.7.1
numpy==1.24.3
pandas==1.5.3
PyYAML==6.0.1
PyYAML==6.0
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.4.1.post1
torch==2.4.0
torch==2.0.1+cu118
torchinfo==1.8.0
torchvision==0.19.0
torchvision==0.15.2+cu118
tqdm==4.66.4
tqdm==4.65.0
~cipy==1.10.1
